Coupled-cluster methods

ElemCo.CoupledCluster — Module

coupled-cluster methods

Exported functions

ElemCo.CoupledCluster.calc_MP2 — Function

calc_MP2(EC::ECInfo)

Calculate closed-shell MP2 energy and amplitudes. Return (EMp2, T2)

ElemCo.CoupledCluster.calc_UMP2 — Function

calc_UMP2(EC::ECInfo, addsingles=true)

Calculate unrestricted MP2 energy and amplitudes. Return (EMp2, T2a, T2b, T2ab)

ElemCo.CoupledCluster.method_name — Function

method_name(T1, dc=false)

Guess method name (CCSD/DCSD/CCD/DCD)

ElemCo.CoupledCluster.calc_cc — Function

calc_cc(EC::ECInfo, T1, T2, dc = false)

Calculate closed-shell coupled cluster amplitudes.

If length(T1) is 0 on input, no singles will be calculated. If dc: calculate distinguishable cluster.

calc_cc(EC::ECInfo, T1a, T1b, T2a, T2b, T2ab, dc = false)

Calculate unrestricted coupled cluster amplitudes.

If length(T1a) && length(T1b) are 0 on input, no singles will be calculated. If dc: calculate distinguishable cluster.

ElemCo.CoupledCluster.calc_pertT — Function

calc_pertT(EC::ECInfo, T1, T2; save_t3=false)

Calculate (T) correction for closed-shell CCSD.

Return ( (T)-energy, [T]-energy))

Some internal functions

ElemCo.CoupledCluster.update_singles — Function

update_singles(R1, ϵo, ϵv, shift)

Calculate update for singles amplitudes.

update_singles(EC::ECInfo, R1; spincase::Symbol=:α, use_shift=true)

Calculate update for singles amplitudes for a given spincase∈{:α,:β}.

ElemCo.CoupledCluster.update_doubles — Function

update_doubles(R2, ϵo1, ϵv1, ϵo2, ϵv2, shift)

Calculate update for doubles amplitudes.

update_doubles(EC::ECInfo, R2; spincase::Symbol=:α, antisymmetrize=false, use_shift=true)

Calculate update for doubles amplitudes for a given spincase∈{:α,:β,:αβ}.

ElemCo.CoupledCluster.calc_singles_energy — Function

calc_singles_energy(EC::ECInfo, T1; fock_only=false)

Calculate coupled-cluster closed-shell singles energy.

calc_singles_energy(EC::ECInfo, T1a, T1b; fock_only=false)

Calculate energy for α (T1a) and β (T1b) singles amplitudes.

ElemCo.CoupledCluster.calc_doubles_energy — Function

calc_doubles_energy(EC::ECInfo, T2; fock_only=false)

Calculate coupled-cluster closed-shell doubles energy.

calc_doubles_energy(EC::ECInfo, T2a, T2b, T2ab; fock_only=false)

Calculate energy for αα (T2a), ββ (T2b) and αβ (T2ab) doubles amplitudes.

ElemCo.CoupledCluster.calc_hylleraas — Function

calc_hylleraas(EC::ECInfo, T1, T2, R1, R2)

Calculate closed-shell singles and doubles Hylleraas energy

calc_hylleraas(EC::ECInfo, T1a, T1b, T2a, T2b, T2ab, R1a, R1b, R2a, R2b, R2ab)

Calculate singles and doubles Hylleraas energy.

ElemCo.CoupledCluster.calc_singles_norm — Function

calc_singles_norm(T1)

Calculate squared norm of closed-shell singles amplitudes.

calc_singles_norm(T1a, T1b)

Calculate squared norm of unrestricted singles amplitudes.

ElemCo.CoupledCluster.calc_doubles_norm — Function

calc_doubles_norm(T2)

Calculate squared norm of closed-shell doubles amplitudes.

calc_doubles_norm(T2a, T2b, T2ab)

Calculate squared norm of unrestricted doubles amplitudes.

ElemCo.CoupledCluster.calc_dressed_ints — Function

calc_dressed_ints(EC::ECInfo, T1, T12, o1::Char, v1::Char, o2::Char, v2::Char)

Dress integrals with singles amplitudes.

The singles and orbspaces for first and second electron are T1, o1, v1 and T12,o2,v2, respectively. The integrals from EC.fd are used and dressed integrals are stored asd_????`.

calc_dressed_ints(EC::ECInfo, T1a, T1b=Float64[])

Dress integrals with singles.

ElemCo.CoupledCluster.calc_D2 — Function

calc_D2(EC::ECInfo, T1, T2, scalepp = false)

Calculate $D^{ij}_{pq} = T^{ij}_{cd} + T^i_c T^j_d +δ_{ik} T^j_d + T^i_c δ_{jl} + δ_{ik} δ_{jl}$. Return as D[pqij]

If scalepp: D[ppij] elements are scaled by 0.5 (for triangular summation).

ElemCo.CoupledCluster.calc_D2a — Function

calc_D2a(EC::ECInfo, T1a, T2a)

Calculate $^{αα}D^{ij}_{pq} = T^{ij}_{cd} + P_{ij}(T^i_c T^j_d +δ_{ik} T^j_d + T^i_c δ_{jl} + δ_{ik} δ_{jl})$ with $P_{ij} X_{ij} = X_{ij} - X_{ji}$. Return as D[pqij]

ElemCo.CoupledCluster.calc_ccsd_resid — Function

calc_ccsd_resid(EC::ECInfo, T1, T2, dc)

Calculate CCSD or DCSD closed-shell residual.

calc_ccsd_resid(EC::ECInfo, T1a,T1b,T2a,T2b,T2ab,dc)

Calculate UCCSD or UDCSD residual.