Generation of integrals using density fitting

ElemCo.DfDump — Module

generate fcidump using df integrals and store in dumpfile

Exported functions

ElemCo.DfDump.dfdump — Function

dfdump(EC::ECInfo)

Generate fcidump using df integrals and store in IntOptions.fcidump. If IntOptions.fcidump is empty, don't write to fcidump file, store in EC.fd.

Internal functions

ElemCo.DfDump.generate_integrals — Function

generate_integrals(EC::ECInfo, fdump::FDump, cMO, full_spaces)

Generate int2, int1 and int0 integrals for fcidump using density fitting.

mp2fit basis is used for int2 integrals, and jkfit basis-correction is used for int1 and int0 integrals. full_spaces is a dictionary with spaces without frozen orbitals.