MCSCF program

Exported functions

ElemCo.DFMCSCF.dfmcscf — Function

dfmcscf(EC::ECInfo; direct = false, guess=:SAD, IterMax=50)

Main body of Density-Fitted Multi-Configurational Self-Consistent-Field method

ElemCo.DFMCSCF.davidson — Function

davidson(H::Matrix, N::Integer, n::Integer, thres::Number, convTrack::Bool=false)

Calculate one of the eigenvalues and corresponding eigenvector of the matrix H (usually the lowest eigenvalue), N is the size of the matrix H, n is the maximal size of projected matrix, thres is the criterion of convergence, convTrack is to decide whether the tracking of eigenvectors is used