FCIDump files

ElemCo.FciDump — Module

Read and write fcidump format integrals. Individual arrays of integrals can also be in *.npy format

The FCIDump file format is a simple text file format for storing molecular integrals. It is used by many quantum chemistry programs.

Storage of molecular integrals

ElemCo.FciDump.FDump — Type

FDump

Molecular integrals

The 2-e integrals are stored in the physicists' notation: int2[pqrs] = <pq|rs> and for triang the last two indices are stored as a single upper triangular index (r <= s)

int2: 2-e⁻ integrals for restricted orbitals fcidump. Default: []

int2aa: αα 2-e⁻ integrals for unrestricted orbitals fcidump. Default: []

int2bb: ββ 2-e⁻ integrals for unrestricted orbitals fcidump. Default: []

int2ab: αβ 2-e⁻ integrals for unrestricted orbitals fcidump. Default: []

int1: 1-e⁻ integrals for restricted orbitals fcidump. Default: []

int1a: α 1-e⁻ integrals for unrestricted orbitals fcidump. Default: []

int1b: β 1-e⁻ integrals for unrestricted orbitals fcidump. Default: []

int0: core energy Default: 0.0

head: header of fcidump file, a dictionary of arrays. Default: Dict()

triang: use an upper triangular index for last two indices of 2e⁻ integrals. Default: true

uhf: a convinience variable, has to coincide with head["IUHF"][1] > 0. Default: false

Exported functions

ElemCo.FciDump.read_fcidump — Function

read_fcidump(fcidump::String)

Read ascii file (possibly with integrals in npy files).

ElemCo.FciDump.write_fcidump — Function

write_fcidump(fd::FDump, fcidump::String, tol=1e-12)

Write fcidump file.

ElemCo.FciDump.fd_exists — Function

fd_exists(fd::FDump)

Return true if the object is a non-empty FDump

ElemCo.FciDump.transform_fcidump — Function

transform_fcidump(fd::FDump, Tl::AbstractArray, Tr::AbstractArray)

Transform integrals to new basis using Tl and Tr transformation matrices. For UHF fcidump, Tl and Tr are arrays of matrices for α and β spin. If Tl and Tr are arrays of arrays, then the function transforms rhf fcidump to uhf fcidump.

ElemCo.FciDump.headvar — Function

headvar(head::Dict, key::String)

Check header for key, return value if a list, or the element or nothing if not there.

headvar(fd::FDump, key::String)

Check header for key, return value if a list, or the element or nothing if not there.

ElemCo.FciDump.integ1 — Function

integ1(fd::FDump, spincase::Symbol=:α)

Return 1-e⁻ integrals (for UHF fcidump: for spincase). spincase can be :α or :β.

ElemCo.FciDump.integ2 — Function

integ2(fd::FDump, spincase::Symbol=:α)

Return 2-e⁻ integrals (for UHF fcidump: for spincase). spincase can be :α, :β or :αβ.

ElemCo.FciDump.uppertriangular — Function

uppertriangular(i1, i2)

Return upper triangular index from two indices i1 <= i2.

uppertriangular(i1, i2, i3)

Return upper triangular index from three indices i1 <= i2 <= i3.

ElemCo.FciDump.uppertriangular_range — Function

uppertriangular_range(i2)

Return range for the upper triangular index (i1 <= i2) for a given i2.