ElemCo.jl global information

ElemCo.ECInfos.ECInfo — Type

ECInfo

Global information for ElemCo.

scr: path to scratch directory. Default: joinpath(tempdir(), "elemcojlscr")

out: output file. Default:

verbosity: verbosity level. Default: 2

options: options. Default: Options()

ms: molecular system. Default: MSys()

fd: fcidump. Default: FDump()

ignore_error: ignore various errors. Default: false

space: subspaces: 'o'ccupied, 'v'irtual, 'O'ccupied-β, 'V'irtual-β, ':' general. Default: Dict{Char, Any}()

nocc: number of occupied orbitals (for UHF: α). Default: 0

noccb: number of occupied orbitals (β). Default: 0

fock: fock matrix (for UHF: α). Default: Float64[]

fockb: fock matrix (β). Default: Float64[]

ϵo: occupied orbital energies (for UHF: α). Default: Float64[]

ϵv: virtual orbital energies (for UHF: α). Default: Float64[]

ϵob: occupied orbital energies (β). Default: Float64[]

ϵvb: virtual orbital energies (β). Default: Float64[]

ElemCo.ECInfos.setup! — Function

setup!(EC::ECInfo; fcidump="", occa="-", occb="-", nelec=0, charge=0, ms2=0)

Setup ECInfo from fcidump or molecular system.

ElemCo.ECInfos.set_options! — Function

set_options!(opt; kwargs...)

Set options for option opt using keyword arguments.

ElemCo.ECInfos.parse_orbstring — Function

parse_orbstring(orbs::String; orbsym = Vector{Int})

Parse a string specifying some list of orbitals, e.g., -3+5-8+10-12 → [1 2 3 5 6 7 8 10 11 12] or use ':' and ';' instead of '-' and '+', respectively.

ElemCo.ECInfos.get_occvirt — Function

get_occvirt(EC::ECInfo, occas::String, occbs::String, norb, nelec; ms2=0, orbsym = Vector{Int})

Use a +/- string to specify the occupation. If occbs=="-", the occupation from occas is used (closed-shell). If both are "-", the occupation is deduced from nelec and ms2. The optional argument orbsym is a vector with length norb of orbital symmetries (1 to 8) for each orbital.