Calculations

Macros

ElemCo.@ECsetup — Macro

@ECsetup()

Setup EC::ECInfo from variables geometry::String and basis::Dict{String,Any}.

Examples

geometry="
He 0.0 0.0 0.0"
basis = Dict("ao"=>"cc-pVDZ", "jkfit"=>"cc-pvtz-jkfit", "mp2fit"=>"cc-pvdz-rifit")
@ECsetup
# output
Occupied orbitals:[1]

ElemCo.@opt — Macro

@opt(what, kwargs...)

Set options for EC::ECInfo.

If EC is not already setup, it will be done.

Examples

@opt scf thr=1.e-14 maxit=10
@opt cc maxit=100

ElemCo.@dfhf — Macro

@dfhf()

Run DFHF calculation and return MO coefficients (ORBS) and orbital energies (EPS).

ElemCo.@dfints — Macro

@dfints(orbs = nothing, fcidump = "")

Generate 2 and 4-idx MO integrals using density fitting.

If orbs::Matrix is given, the orbitals are used to generate the integrals, otherwise the last orbitals (ORBS) are used. If fcidump::String is given, the integrals are written to the fcidump file.

ElemCo.@cc — Macro

@cc(method, kwargs...)

Run coupled cluster calculation.

The type of the method is determined by the first argument (ccsd/ccsd(t)/dcsd etc)

Keyword arguments

fcidump::String: fcidump file (default: "", i.e., use integrals from EC).

occa::String: occupied α orbitals (default: "-").

occb::String: occupied β orbitals (default: "-").

Examples

geometry="bohr
O      0.000000000    0.000000000   -0.130186067
H1     0.000000000    1.489124508    1.033245507
H2     0.000000000   -1.489124508    1.033245507"
basis = Dict("ao"=>"cc-pVDZ", "jkfit"=>"cc-pvtz-jkfit", "mp2fit"=>"cc-pvdz-rifit")
@dfhf
@dfints
@cc ccsd

Exported functions

ElemCo.ECdriver — Function

ECdriver(EC::ECInfo, methods; fcidump="FCIDUMP", occa="-", occb="-")

Run electronic structure calculation for EC::ECInfo using methods methods::String.

The integrals are read from fcidump::String (default: "FCIDUMP"). If fcidump::String is empty, the integrals from EC.fd are used. The occupied α orbitals are given by occa::String (default: "-"). The occupied β orbitals are given by occb::String (default: "-"). If occb::String is empty, the occupied β orbitals are the same as the occupied α orbitals (closed-shell case).

Some internal functions

ElemCo.parse_commandline — Function

parse_commandline(EC::ECInfo)

Parse command line arguments.

ElemCo.is_closed_shell — Function

is_closed_shell(EC::ECInfo)

Check if the system is closed-shell according the to the reference occupation and FCIDump.

ElemCo.calc_HF_energy — Function

calc_HF_energy(EC::ECInfo, closed_shell)

Calculate HF energy from FCIDump and EC info.