Molecular system

The molecular system is the core of the simulation. It contains all the information about the molecule, including the geometry and basis sets The molecular system is defined using the MSys function:

MSys(geometry, basis)

where geometry is a string containing the molecular geometry in the XYZ format, and basis is a dictionary containing the basis set information.

Geometry

The geometry of the molecule is defined using the geometry argument of the MSys function. The geometry is defined in the XYZ format. Here's an example of how you can define the geometry of a water molecule:

geometry="bohr
     O      0.000000000    0.000000000   -0.130186067
     H1     0.000000000    1.489124508    1.033245507
     H2     0.000000000   -1.489124508    1.033245507"

The first line of the geometry string contains the units of the coordinates. The supported units are bohr and angstrom. The coordinates of the atoms are specified in the following lines. Each line contains the atomic symbol and the coordinates of the atom. The coordinates are separated by spaces or tabs.

Basis set

The basis set is defined using the basis argument of the MSys function. The basis set is defined as a dictionary, where the keys are the names of the basis sets, and the values are the basis set definitions. Here's an example of how you can define the basis set for a water molecule:

basis = Dict("ao"=>"cc-pVDZ",
             "jkfit"=>"cc-pvtz-jkfit",
             "mp2fit"=>"cc-pvdz-rifit")

The basis set dictionary contains three keys: ao, jkfit, and mp2fit. The ao key contains the basis set for the AO integrals, the jkfit key contains the basis set for the density fitting integrals in the Hartree-Fock calculations, and the mp2fit key contains the fitting basis set for the correlated calculations.

Exported functions and types

ElemCo.MSystem.MSys — Type

Geometry and basis set for each element name in the geometry

atoms: array of atomic centers.

ElemCo.MSystem.ms_exists — Function

ms_exists(ms::MSys)

Check whether the system is not empty.

ElemCo.MSystem.Basis — Type

Basis set

name: e.g., vdz, cc-pvdz, aug-cc-pvdz, cc-pvdz-jkfit.

ElemCo.MSystem.ACenter — Type

Atomic center

name: name (as defined in input).

nuccharge: nuclear charge (can be changed...).

coord: coordinates in bohr.

basis: various basis sets (ao,mp2fit,jkfit).

ElemCo.MSystem.genxyz — Function

genxyz(ac::ACenter; bohr=true)

Generate xyz string with element without numbers. If bohr: return in coordinates in bohr.

genxyz(ms::MSys; bohr=true)

Generate xyz string with elements without numbers. If bohr: return in coordinates in bohr

ElemCo.MSystem.nuclear_repulsion — Function

nuclear_repulsion(ms::MSys)

Calculate nuclear repulsion energy.

ElemCo.MSystem.bond_length — Function

bond_length(cen1::ACenter, cen2::ACenter)

Calculate bond length between two centers.

ElemCo.MSystem.electron_distribution — Function

electron_distribution(elem::AbstractString, nsh4l::Vector{Int})

Distribute electrons among first atomic orbitals in nsh4l[1]s nsh4l[2]p nsh4l[3]d nsh4l[4]f... order considering the Hund's rule and electron configuration of the atom. Average occupations to account for the spin degeneracy and hybridization.

electron_distribution(elnam::AbstractString, minbas::AbstractString)

Return the averaged number of electrons in the orbitals in the minimal basis set.

Number of orbitals in the minimal basis set has to be specified in minbas.jl.

ElemCo.MSystem.guess_nelec — Function

guess_nelec(ms::MSys)

Guess the number of electrons in the neutral system.

ElemCo.MSystem.guess_norb — Function

guess_norb(ms::MSys)

Guess the number of orbitals in the system.