Options

ElemCo.ECInfos.Options — Type

Options for ElemCo.jl.

wf: Wavefunction options. Default: WfOptions()

scf: SCF options. Default: ScfOptions()

int: Integral options. Default: IntOptions()

cc: Coupled-Cluster options. Default: CcOptions()

cholesky: Cholesky options. Default: CholeskyOptions()

diis: DIIS options. Default: DiisOptions()

ElemCo.ECInfos.WfOptions — Type

Options for wavefunction/orbitals.

ms2: spin magnetic quantum number times two (2×mₛ) of the system. Default: -1

nelec: number of electrons. If < 0, the number of electrons is read from the FCIDump file or guessed for the neutral system. Default: -1

charge: charge of the system (relative to nelec/FCIDump/neutral system!). Default: 0

orb: filename of MO coefficients. Used by all programs to read and write orbitals from/to file. Default: C_Am

left: addition to the filename for left orbitals (for biorthogonal calculations). Default: -left

core: core type for frozen-core approximation:

:none (no frozen-core approximation),

:small (semi-core orbitals correlated),

:large (semi-core orbitals frozen). Default: :large

freeze_nocc: number of occupied (core) orbitals to freeze (overwrites core). Default: -1

freeze_nvirt: number of virtual (highest) orbitals to freeze. Default: 0

occa: occupied α (or closed-shell) orbitals. Default: -

occb: occupied β orbitals. If occb::String is empty, the occupied β orbitals are the same as the occupied α orbitals (closed-shell case). Default: -

ElemCo.ECInfos.ScfOptions — Type

Options for SCF calculation.

thr: convergence threshold. Default: 1.0e-10

maxit: maximum number of iterations. Default: 50

imagtol: tolerance for imaginary part of MO coefs (for biorthogonal). Default: 1.0e-8

direct: direct calculation without storing integrals. Default: false

guess: orbital guess. Default: :SAD

ElemCo.ECInfos.CcOptions — Type

Options for Coupled-Cluster calculation.

thr: convergence threshold. Default: 1.0e-10

maxit: maximum number of iterations. Default: 50

shifts: level shift for singles. Default: 0.15

shiftp: level shift for doubles. Default: 0.2

shiftt: level shift for triples. Default: 0.2

ampsvdtol: amplitude decomposition threshold. Default: 0.001

use_kext: use kext for doubles residual. Default: true

calc_d_vvvv: calculate dressed <vv|vv>. Default: false

calc_d_vvvo: calculate dressed <vv|vo>. Default: false

calc_d_vovv: calculate dressed <vo|vv>. Default: false

calc_d_vvoo: calculate dressed <vv|oo>. Default: false

triangular_kext: use a triangular kext if possible. Default: true

calc_t3_for_decomposition: calculate (T) for decomposition. Default: false

deco_ishiftp: imaginary shift for denominator in doubles decomposition. Default: 0.0

deco_ishiftt: imaginary shift for denominator in triples decomposition. Default: 0.0

use_projx: use a projected exchange for contravariant doubles amplitudes in SVD-DCSD, $\tilde T_{XY} = U^{†a}_{iX} U^{†b}_{jY} \tilde T^{ij}_{ab}$. Default: false

use_full_t2: use full doubles amplitudes in SVD-DCSD. The decomposition is used only for $N^6$ scaling terms. Default: false

project_vovo_t2: what to project in $v_{ak}^{ci} T^{kj}_{cb}$ in SVD-DCSD: 0: both, 1: amplitudes, 2: residual, 3: robust fit. Default: 2

decompose_full_doubles: decompose full doubles amplitudes in SVD-DCSD (slow). Default: false

start: main part of filename for start amplitudes. For example, the singles amplitudes are read from start*"_singles" Default: cc_amplitudes

save: main part of filename to save amplitudes. For example, the singles amplitudes are saved to save*"_singles" Default: cc_amplitudes

ElemCo.ECInfos.CholeskyOptions — Type

Options for Cholesky decomposition.

thr: cholesky threshold. Default: 1.0e-6

ElemCo.ECInfos.DiisOptions — Type

Options for DIIS.

maxdiis: maximum number of DIIS vectors. Default: 6

resthr: DIIS residual threshold. Default: 10.0