Tensor tools

ElemCo.TensorTools — Module

tensor tools, e.g., access to integrals, load/save intermediates...

I/O functions

ElemCo.TensorTools.save — Function

save(EC::ECInfo, fname::String, a::AbstractArray, descr="tmp"; overwrite=true)

Save array a to file fname in EC.scr directory. Add file to EC.files with description descr.

ElemCo.TensorTools.load — Function

load(EC::ECInfo, fname::String)

Load array from file fname in EC.scr directory.

ElemCo.TensorTools.mmap — Function

mmap(EC::ECInfo, fname::String)

Memory-map an existing file for reading. Return a pointer to the file and the mmaped array.

ElemCo.TensorTools.newmmap — Function

newmmap(EC::ECInfo, fname::String, Type, dims::Tuple{Vararg{Int}}, descr="tmp")

Create a new memory-map file for writing (overwrites existing file). Add file to EC.files with description descr. Return a pointer to the file and the mmaped array.

ElemCo.TensorTools.closemmap — Function

closemmap(EC::ECInfo, file, array)

Close memory-map file and flush to disk.

Integral extraction

ElemCo.TensorTools.ints1 — Function

ints1(EC::ECInfo, spaces::String, spincase = nothing)

Return subset of 1e⁻ integrals according to spaces.

The spincase∈{:α,:β} can explicitly be given, or will be deduced from upper/lower case of spaces specification.

ElemCo.TensorTools.ints2 — Function

ints2(EC::ECInfo, spaces::String, spincase = nothing, detri = true)

Return subset of 2e⁻ integrals according to spaces.

The spincase∈{:α,:β} can explicitly be given, or will be deduced from upper/lower case of spaces specification. If the last two indices are stored as triangular and detri - make them full, otherwise return as a triangular cut.

Tensor manipulation

ElemCo.TensorTools.sqrtinvchol — Function

sqrtinvchol(A::AbstractMatrix; tol = 1e-8, verbose = false)

Return NON-SYMMETRIC (pseudo)sqrt-inverse of a hermitian matrix using Cholesky decomposition.

Starting from A^{-1} = A^{-1} L (A^{-1} L)† = M M† with A = L L†. By solving the equation L† M = I (for low-rank: using QR decomposition). Return M

ElemCo.TensorTools.invchol — Function

invchol(A::AbstractMatrix; tol = 1e-8, verbose = false)

Return (pseudo)inverse of a hermitian matrix using Cholesky decomposition .

A^-1 = A^-1 L (A^-1 L)† = M M† with A = L L†. By solving the equation L† M = I (for low-rank: using QR decomposition)

ElemCo.TensorTools.rotate_eigenvectors_to_real! — Function

rotate_eigenvectors_to_real!(evecs::AbstractMatrix, evals::AbstractVector)

In-place transform complex eigenvectors of a real matrix to a real space such that they block-diagonalize the matrix.