Coupled-cluster methods

ElemCo.CoupledCluster — Module

Coupled-cluster methods

Exported functions

ElemCo.CoupledCluster.calc_MP2 — Function

calc_MP2(EC::ECInfo, addsingles=true)

Calculate closed-shell MP2 energy and amplitudes. The amplitudes are stored in T_vvoo file. If addsingles: singles are also calculated and stored in T_vo file. Return EMp2

ElemCo.CoupledCluster.calc_UMP2 — Function

calc_UMP2(EC::ECInfo, addsingles=true)

Calculate unrestricted MP2 energy and amplitudes. The amplitudes are stored in T_vvoo, T_VVOO, and T_vVoO files. If addsingles: singles are also calculated and stored in T_vo and T_VO files. Return EMp2

ElemCo.CoupledCluster.calc_cc — Method

calc_cc(EC::ECInfo, method::ECMethod)

Calculate coupled cluster amplitudes.

Exact specification of the method is given by method.

ElemCo.CoupledCluster.calc_pertT — Method

calc_pertT(EC::ECInfo; save_t3=false)

Calculate (T) correction for closed-shell CCSD.

Return ( (T)-energy, [T]-energy))

Internal functions

ElemCo.CoupledCluster.add_to_singles_and_doubles_residuals — Method

add_to_singles_and_doubles_residuals(EC, R1, R2)

Add contributions from triples to singles and doubles residuals.

ElemCo.CoupledCluster.calc_4idx_T3T3_XY — Method

calc_4idx_T3T3_XY(EC::ECInfo, T2, UvoX, ϵX)

Calculate $D^{ij}_{ab} = T^i_{aXY} T^j_{bXY}$ using half-decomposed imaginary-shifted perturbative triple amplitudes $T^i_{aXY}$ from T2 (and UvoX)

ElemCo.CoupledCluster.calc_D2 — Function

calc_D2(EC::ECInfo, T1, T2, scalepp=false)

Calculate $D^{ij}_{pq} = T^{ij}_{cd} + T^i_c T^j_d +δ_{ik} T^j_d + T^i_c δ_{jl} + δ_{ik} δ_{jl}$. Return as D[pqij]

If scalepp: D[ppij] elements are scaled by 0.5 (for triangular summation).

ElemCo.CoupledCluster.calc_D2a — Method

calc_D2a(EC::ECInfo, T1a, T2a)

Calculate $^{αα}D^{ij}_{pq} = T^{ij}_{cd} + P_{ij}(T^i_c T^j_d +δ_{ik} T^j_d + T^i_c δ_{jl} + δ_{ik} δ_{jl})$ with $P_{ij} X_{ij} = X_{ij} - X_{ji}$. Return as D[pqij]

ElemCo.CoupledCluster.calc_D2ab — Function

calc_D2ab(EC::ECInfo, T1a, T1b, T2ab, scalepp=false)

Calculate $^{αβ}D^{ij}_{pq} = T^{ij}_{cd} + T^i_c T^j_d +δ_{ik} T^j_d + T^i_c δ_{jl} + δ_{ik} δ_{jl}$ Return as D[pqij]

If scalepp: D[ppij] elements are scaled by 0.5 (for triangular summation)

ElemCo.CoupledCluster.calc_D2b — Method

calc_D2b(EC::ECInfo, T1b, T2b)

Calculate $^{ββ}D^{ij}_{pq} = T^{ij}_{cd} + P_{ij}(T^i_c T^j_d +δ_{ik} T^j_d + T^i_c δ_{jl} + δ_{ik} δ_{jl})$ with $P_{ij} X_{ij} = X_{ij} - X_{ji}$. Return as D[pqij]

ElemCo.CoupledCluster.calc_ccsd_resid — Method

calc_ccsd_resid(EC::ECInfo, T1, T2, dc)

Calculate CCSD or DCSD closed-shell residual.

ElemCo.CoupledCluster.calc_ccsd_resid — Method

calc_ccsd_resid(EC::ECInfo, T1a, T1b, T2a, T2b, T2ab; dc=false, tworef=false, fixref=false)

Calculate UCCSD or UDCSD residual.

ElemCo.CoupledCluster.calc_ccsdt — Function

calc_ccsdt(EC::ECInfo, useT3=false, cc3=false)

Calculate decomposed closed-shell DC-CCSDT amplitudes.

If useT3: (T) amplitudes from a preceding calculations will be used as starting guess. If cc3: calculate CC3 amplitudes.

ElemCo.CoupledCluster.calc_doubles_energy — Method

calc_doubles_energy(EC::ECInfo, T2a, T2b, T2ab; fock_only=false)

Calculate energy for αα (T2a), ββ (T2b) and αβ (T2ab) doubles amplitudes.

ElemCo.CoupledCluster.calc_doubles_energy — Method

calc_doubles_energy(EC::ECInfo, T2; fock_only=false)

Calculate coupled-cluster closed-shell doubles energy.

ElemCo.CoupledCluster.calc_dressed_ints — Function

calc_dressed_ints(EC::ECInfo, T1a, T1b=Float64[])

Dress integrals with singles.

ElemCo.CoupledCluster.calc_dressed_ints — Method

calc_dressed_ints(EC::ECInfo, T1, T12, o1::Char, v1::Char, o2::Char, v2::Char)

Dress integrals with singles amplitudes.

The singles and orbspaces for first and second electron are T1, o1, v1 and T12, o2, v2, respectively. The integrals from EC.fd are used and dressed integrals are stored as d_????.

ElemCo.CoupledCluster.calc_hylleraas — Method

calc_hylleraas(EC::ECInfo, T1a, T1b, T2a, T2b, T2ab, R1a, R1b, R2a, R2b, R2ab)

Calculate singles and doubles Hylleraas energy.

ElemCo.CoupledCluster.calc_hylleraas — Method

calc_hylleraas(EC::ECInfo, T1, T2, R1, R2)

Calculate closed-shell singles and doubles Hylleraas energy

ElemCo.CoupledCluster.calc_hylleraas4spincase — Method

calc_hylleraas4spincase(EC::ECInfo, o1, v1, o2, v2, T1, T2, R1, R2, fov)

Calculate singles and doubles Hylleraas energy for one spin case.

ElemCo.CoupledCluster.calc_singles_energy — Method

calc_singles_energy(EC::ECInfo, T1a, T1b; fock_only=false)

Calculate energy for α (T1a) and β (T1b) singles amplitudes.

ElemCo.CoupledCluster.calc_singles_energy — Method

calc_singles_energy(EC::ECInfo, T1; fock_only=false)

Calculate coupled-cluster closed-shell singles energy.

ElemCo.CoupledCluster.calc_triples_decomposition — Method

calc_triples_decomposition(EC::ECInfo)

Decompose $T^{ijk}_{abc}$ as $U^{iX}_a U^{jY}_b U^{kZ}_c T_{XYZ}$.

ElemCo.CoupledCluster.calc_triples_decomposition_without_triples — Method

calc_triples_decomposition_without_triples(EC::ECInfo, T2)

Decompose $T^{ijk}_{abc}$ as $U^{iX}_a U^{jY}_b U^{kZ}_c T_{XYZ}$ without explicit calculation of $T^{ijk}_{abc}$.

Compute perturbative $T^i_{aXY}$ and decompose $D^{ij}_{ab} = (T^i_{aXY} T^j_{bXY})$ to get $U^{iX}_a$.

ElemCo.CoupledCluster.calc_triples_residuals — Function

calc_triples_residuals(EC::ECInfo, T1, T2, cc3 = false)

Calculate decomposed triples DC-CCSDT or CC3 residuals.

ElemCo.CoupledCluster.dress_fock_closedshell — Method

dress_fock_closedshell(EC::ECInfo, T1)

Dress the fock matrix (closed-shell). The dressed fock matrix is stored as df_mm.

ElemCo.CoupledCluster.dress_fock_oppositespin — Method

dress_fock_oppositespin(EC::ECInfo)

Add the dressed opposite-spin part to the dressed Fock matrix.

ElemCo.CoupledCluster.dress_fock_samespin — Method

dress_fock_samespin(EC::ECInfo, T1, o1::Char, v1::Char)

Dress the fock matrix (same-spin part).

ElemCo.CoupledCluster.pseudo_dressed_ints — Function

pseudo_dressed_ints(EC::ECInfo, unrestricted=false)

Save non-dressed integrals in files instead of dressed integrals.

ElemCo.CoupledCluster.test_UaiX — Method

test_UaiX(EC, UaiX)

Test UaiX

ElemCo.CoupledCluster.test_add_to_singles_and_doubles_residuals — Method

test_add_to_singles_and_doubles_residuals(R1, R2, T1, T2)

Test R1(T3) and R2(T3)

ElemCo.CoupledCluster.test_calc_pertT_from_T3 — Method

test_calc_pertT_from_T3(EC, T3)

Test [T]

ElemCo.CoupledCluster.test_dressed_ints — Method

test_dressed_ints(EC, T1)

Compare 3-idx dressed integrals to 4-idx dressed integrals.