ElemCo.jl global information

ElemCo.ECInfos — Module

Various global infos

ElemCo.ECInfos.ECInfo — Type

ECInfo

Global information for ElemCo.

scr: path to scratch directory. Default: joinpath(tempdir(), "elemcojlscr")

ext: extension of temporary files. Default: .bin

out: output file. Default:

verbosity: verbosity level. Default: 2

options: options. Default: Options()

ms: molecular system. Default: MSys()

fd: fcidump. Default: FDump()

files: information about (temporary) files. The naming convention is: prefix_ + name (+extension EC.ext added automatically). prefix can be:

d for dressed integrals

S for overlap matrix

f for Fock matrix

e for orbital energies

D for density matrix

h for core Hamiltonian

C for transformation from one basis to another

name is given by the subspaces involved:

o for occupied

v for virtual

O for occupied-β

V for virtual-β

m for (full) MO space

M for (full) β-MO space

A for AO basis

a for active orbitals

c for closed-shell (doubly-occupied) orbitals

P for auxiliary orbitals (fitting basis)

L for auxiliary orbitals (Cholesky decomposition, orthogonal)

X for auxiliary orbitals (amplitudes decomposition)

Default: Dict{String, String}()

ignore_error: ignore various errors. Default: false

space: subspaces: 'o'ccupied, 'v'irtual, 'O'ccupied-β, 'V'irtual-β, ':' general. Default: Dict{Char, Any}()

ElemCo.ECInfos.setup! — Function

setup!(EC::ECInfo; fcidump="", occa="-", occb="-", nelec=0, charge=0, ms2=0)

Setup ECInfo from fcidump or molecular system.

ElemCo.ECInfos.set_options! — Function

set_options!(opt; kwargs...)

Set options for option opt using keyword arguments.

ElemCo.ECInfos.parse_orbstring — Function

parse_orbstring(orbs::String; orbsym = Vector{Int})

Parse a string specifying some list of orbitals, e.g., -3+5-8+10-12 → [1 2 3 5 6 7 8 10 11 12] or use ':' and ';' instead of '-' and '+', respectively.

ElemCo.ECInfos.get_occvirt — Function

get_occvirt(EC::ECInfo, occas::String, occbs::String, norb, nelec; ms2=0, orbsym = Vector{Int})

Use a +/- string to specify the occupation. If occbs=="-", the occupation from occas is used (closed-shell). If both are "-", the occupation is deduced from nelec and ms2. The optional argument orbsym is a vector with length norb of orbital symmetries (1 to 8) for each orbital.

ElemCo.ECInfos.n_occ_orbs — Function

n_occ_orbs(EC::ECInfo)

Return number of occupied orbitals (for UHF: α).

ElemCo.ECInfos.n_occb_orbs — Function

n_occb_orbs(EC::ECInfo)

Return number of occupied orbitals (β).

ElemCo.ECInfos.n_virt_orbs — Function

n_virt_orbs(EC::ECInfo)

Return number of virtual orbitals (for UHF: α).

ElemCo.ECInfos.n_virtb_orbs — Function

n_virtb_orbs(EC::ECInfo)

Return number of virtual orbitals (β).

ElemCo.ECInfos.n_orbs — Function

n_orbs(EC::ECInfo)

Return number of orbitals.

File management

ElemCo.ECInfos.file_exists — Function

file_exists(EC::ECInfo, name::String)

Check if file name exists in ECInfo.

ElemCo.ECInfos.add_file — Function

add_file(EC::ECInfo, name::String, descr::String; overwrite = false)

Add file name to ECInfo with (space-separated) descriptions descr. Possible description: tmp (temporary).

ElemCo.ECInfos.delete_temporary_files — Function

delete_temporary_files(EC::ECInfo)

Delete all temporary files in ECInfo.

Abstract types

ElemCo.AbstractEC — Module

Abstract types to resolve circular dependencies

Internal functions

ElemCo.ECInfos.symorb2orb — Function

symorb2orb(symorb::SubString, symoffset::Vector{Int})

Convert a symorb (like 1.3 [orb.sym]) to an orbital number. If no sym given, just return the orbital number converted to Int.